migrated new functionality to ResourcesListMobile, tags alphabetized
